video_core: Commonize shader uniform data
